Move schema version endpoint to schema controller. Migrate related service and repository logic to schema service and repository
