+++ src/main/java/org/zalando/nakadi/service/SchemaService.java
@@ -1,8 +1,13 @@
 package org.zalando.nakadi.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.zalando.nakadi.domain.EventTypeSchema;
+import org.zalando.nakadi.exceptions.IllegalVersionNumberException;
+import org.zalando.nakadi.exceptions.InternalNakadiException;
+import org.zalando.nakadi.exceptions.NoSuchSchemaException;
 import org.zalando.nakadi.repository.db.SchemaRepository;
 import org.zalando.problem.Problem;
 
@@ -11,6 +16,8 @@ import java.util.List;
 
 @Component
 public class SchemaService {
+
+    private static final Logger LOG = LoggerFactory.getLogger(SchemaService.class);
 
     private final SchemaRepository schemaRepository;
     private final PaginationService paginationService;
@@ -37,4 +44,19 @@ public class SchemaService {
                 .paginate(schemas, offset,  limit, "/schemas", () -> schemaRepository.getSchemasCount(name)));
     }
 
+    public Result<EventTypeSchema> getSchemaVersion(String name, String version) {
+        try {
+            final EventTypeSchema schema = schemaRepository.getSchemaVersion(name, version);
+            return Result.ok(schema);
+        } catch (final IllegalVersionNumberException e) {
+            LOG.debug("Malformed version number: {}", version);
+            return Result.problem(e.asProblem());
+        } catch (final NoSuchSchemaException e) {
+            LOG.debug("Could not find EventTypeSchema version: {} for EventType: {}", version, name);
+            return Result.problem(e.asProblem());
+        } catch (final InternalNakadiException e) {
+            LOG.error("Problem loading event type schema version " + version + " for EventType " + name, e);
+            return Result.problem(e.asProblem());
+        }
+    }
 }
